+++ src/main/java/com/sparta/bipuminbe/requests/service/SupplyRequestService.java
@@ -63,13 +63,9 @@ public class SupplyRequestService {
         checkSupplyRequest(request);
 
         AcceptResult acceptResult = AcceptResult.valueOf(supplyProcessRequestDto.getAcceptResult());
-        checkAcceptResult(acceptResult);
-
-        String comment = supplyProcessRequestDto.getComment();
-        request.processingRequest(acceptResult, comment);
+        request.processingRequest(acceptResult, supplyProcessRequestDto.getComment());
 
         if (acceptResult.equals(AcceptResult.DECLINE)) {
-            checkNullComment(comment);
             return ResponseDto.success("승인 거부 완료.");
         }
 
@@ -82,20 +78,6 @@ public class SupplyRequestService {
     // 비품요청의 승인에는 비품이 필요하다.
     private void checkNullSupplyId(Long supplyId) {
         if (supplyId == null) {
-            throw new CustomException(ErrorCode.NotAllowedMethod);
-        }
-    }
-
-    // 거절시 거절 사유 작성은 필수다.
-    private void checkNullComment(String comment) {
-        if (comment == null || comment.equals("")) {
-            throw new CustomException(ErrorCode.NullComment);
-        }
-    }
-
-    // 폐기는 수리 요청에만 존재해야 한다.
-    private void checkAcceptResult(AcceptResult acceptResult) {
-        if (acceptResult.equals(AcceptResult.DISPOSE)) {
             throw new CustomException(ErrorCode.NotAllowedMethod);
         }
     }
